+++ synoldap/appinfo/routes.php
@@ -12,6 +12,8 @@ return [
         // API DSM Synology
         ['name' => 'admin#testDsmApi',      'url' => '/admin/test-dsm-api',     'verb' => 'POST'],
         ['name' => 'admin#testSmb',          'url' => '/admin/test-smb',         'verb' => 'POST'],
+        ['name' => 'admin#getUserLdapStatus','url' => '/admin/user-ldap-status', 'verb' => 'GET'],
+        ['name' => 'admin#syncUserLdap',     'url' => '/admin/sync-user-ldap',   'verb' => 'POST'],
         ['name' => 'admin#discoverAcl',     'url' => '/admin/discover-acl',     'verb' => 'GET'],
         ['name' => 'admin#clearAclCache',   'url' => '/admin/clear-acl-cache',  'verb' => 'POST'],
     ],